Move TraceJSONEncoder to JSON helpers
