+++ homeassistant/components/websocket_api/commands.py
@@ -7,7 +7,6 @@ import voluptuous as vol
 from homeassistant.auth.permissions.const import CAT_ENTITIES, POLICY_READ
 from homeassistant.bootstrap import SIGNAL_BOOTSTRAP_INTEGRATONS
 from homeassistant.components.websocket_api.const import ERR_NOT_FOUND
-from homeassistant.components.websocket_api.utils import SubscribeTriggerJSONEncoder
 from homeassistant.const import EVENT_STATE_CHANGED, EVENT_TIME_CHANGED, MATCH_ALL
 from homeassistant.core import DOMAIN as HASS_DOMAIN, callback
 from homeassistant.exceptions import (
@@ -19,6 +18,7 @@ from homeassistant.exceptions import (
 from homeassistant.helpers import config_validation as cv, entity, template
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.event import TrackTemplate, async_track_template_result
+from homeassistant.helpers.json import TraceJSONEncoder
 from homeassistant.helpers.service import async_get_all_descriptions
 from homeassistant.loader import IntegrationNotFound, async_get_integration
 from homeassistant.setup import DATA_SETUP_TIME, async_get_loaded_integrations
@@ -423,7 +423,7 @@ async def handle_subscribe_trigger(hass, connection, msg):
             msg["id"], {"variables": variables, "context": context}
         )
         connection.send_message(
-            json.dumps(message, cls=SubscribeTriggerJSONEncoder, allow_nan=False)
+            json.dumps(message, cls=TraceJSONEncoder, allow_nan=False)
         )
 
     connection.subscriptions[msg["id"]] = (
